fix: pre-production audit sweep — 150+ CRITICAL/HIGH/MEDIUM bugs across 41 sections
Parallel multi-agent sweep against criticalaudit.md. Every bug in Part 1
(sections 1-41) addressed. Part 2 (sections 42-54, UX enrichment) left
as product backlog.

Highlights:

Counters + migrations (072-085)
- Atomic counters table replaces MAX+1 ID allocation for tickets,
  invoices, credit notes, POs, SKUs, BlockChyp refs. Kills both
  poisoning-by-negative-ID and concurrent insert races (I4-I7).
- Unique constraints on ticket order_id + stripe webhook events.
- Password history + idle-session last_active column (073).
- customer_id nullable on tickets/estimates for GDPR erasure (074).
- POS payments: processor/reference/employee_tips (075).
- Inventory cost_locked flag (076), estimate token expiry (077),
  catalog price history (079), payment idempotency + signature
  tracking (080), automation run log (081), webhook delivery
  failures (082), setup tokens (083), import locks + rate limits
  (084), upload quotas (085).

Shared utilities
- New utils/counters.ts with allocateCounter atomic allocator +
  format helpers.
- Extended utils/validate.ts: validateEmail, validatePhoneDigits,
  validateIsoDate, validateEnum, validateHexColor, validateJsonPayload,
  validateArrayBounds, validatePositiveAmount, validateSignedAmount,
  validateIntegerQuantity, validateRequiredString, roundCents/toCents/fromCents.
- New utils/csrf.ts (double-submit cookie) + utils/fileValidation.ts
  (magic-byte MIME check + ClamAV hook) + middleware/fileUploadValidator.ts.

Auth / session hardening
- Password reset column fix (password_hash not password) — every reset
  was silently broken. Reset now revokes sessions + logs history.
- POST /account/2fa/disable, /force-disable-2fa/:id (admin),
  /recover-with-backup-code endpoints.
- Login constant-time path (username OR email, same timing).
- Concurrent-session cap (5), idle-session timeout (14 d),
  jwt.sign + verify pinned to HS256 + issuer + audience.
- Switch-user requires target's 2FA when enabled.
- POST /auth/change-password + /change-pin endpoints (Android).
- Device-trust cookie bound to UA+IP fingerprint.
- Generic "Invalid credentials" 401 everywhere (kills enumeration).

Money + POS + inventory
- POS unit_price forced from inventory_items.retail_price (price spoof killed).
- Atomic guarded stock decrement (WHERE in_stock >= ?) in POS + tickets.
- Split payments + line items + stock all in one transaction.
- Walk-in customer sentinel row, membership discount applied server-side.
- Refund cap against amount_paid - prior_refunds, card-vs-cash routing,
  commissions reversed on approve, SC1 MAX() scalar two-step fix.
- Credit notes: store-credit overflow, never negative amount_due.
- Invoice void restores stock regardless of ticket_id.
- PO receive loop wrapped in transaction.
- Kit component decrement helper exported for POS.

Input validation
- Customers, leads, invoices, estimates, inventory, settings,
  catalog, POS, tickets all validated at edges. Email/phone/date/
  enum/hex-color/JSON/array-bounds applied where the audit flagged
  V1-V24.
- XSS vector on status color (V22) killed.

SQL injection + multi-tenant
- repairDeskImport.ts whitelist on table + trigger names, LIKE
  wildcard escaping (Q1-Q5).
- Super-admin + master-admin dynamic UPDATE whitelists.
- Automations tenant-isolation comment + audit trail.
- Tenant resolver: strict Host header + X-Forwarded-Host proxy check,
  404 for both not-found and suspended tenants.
- WebSocket origin allowlist (WS1).
- reimport-notes.ts + migrate-all-tenants.ts tenant-scoped with
  per-tenant timeout + failed_tenants surface.

Billing
- Stripe: event-age check (<300s), idempotency via PK + OR IGNORE,
  payment_failed grace (3 strikes) + downgrade + notification queue,
  concurrent checkout lock, Stripe-level idempotency keys, updateSubscription
  helper wired into super-admin plan change with rollback on failure.
- BlockChyp: unique transaction refs via counter + 8-byte random,
  per-invoice idempotency via (invoice_id, client_request_id),
  signature file tracking + cleanup on void, tip-adjust endpoint stub,
  test-mode mid-flight guard.

Backups + tenant lifecycle
- Restore endpoint (integrity check + safety copy + per-tenant lock).
- Dedicated BACKUP_ENCRYPTION_KEY, versioned file header, disk-space
  check, per-tenant mutex, ms-precision timestamps.
- Tenant DB files are sacred: deletion archives with -archived-{iso}
  suffix. Slug reclaim closed. 30-day grace period + exportTenantData
  helper. Setup tokens hashed with expiry + consumed_at.

Frontend + Android
- Web: single-flight token refresh, InventoryDetailPage unwrap fix,
  ProtectedRoute loading gate, rememberSaveable-like useEffect cleanup,
  DOMPurify on receipt/invoice footer/header/terms, signature max-size,
  CommandPalette cleanup, LOGOUT_REQUIRED_EVENT, setup-token abort guard.
- Android: OfflineIdGenerator (positive monotonic temp IDs),
  Room money Double->Long cents via Migration 2->3, @ForeignKey cascades,
  clearUserData() on logout, SQLCipher integration plan + DatabaseModule
  wiring, cert pinning (release) + hostname-restricted dev trust,
  HMAC-signed base URL, sync queue dead-letter, sync isolation per entity,
  CameraX gallery picker replaces fake counter, duplicate scanner screen
  deleted, change password/PIN dialogs, per-section dashboard errors.

Reports + misc
- reports.routes.ts: real COGS / margin / tax (killed hardcoded 100%
  margin lie), export-all path, LEFT JOIN instead of dropping rows.
- catalog scraper: EU price format, selector fallback chain, SHA-256
  external IDs, price history respects cost_locked.
- Shared statuses: notify_customer set per status + new setup wizard
  step StepDefaultStatuses to customize before first ticket.
- index.ts: apiRateMap LRU, HTTP->HTTPS redirect sanitization, CORS
  no-Origin rejection in production, WS origin allowlist, membership
  renewal serial loop, cron interval tracking + shutdown clearInterval,
  audit retention 730 days, per-tenant cron tz.
- Electron mgmt: localhost-only rejectUnauthorized, sandbox: true,
  trusted update path resolution, spawn with args arrays, update
  rollback snapshot, IPC stubs return NOT_IMPLEMENTED instead of
  lying success.

README + criticalaudit.md updated with full status.

+++ packages/android/app/src/main/java/com/bizarreelectronics/crm/data/local/db/entities/ExpenseEntity.kt
@@ -5,6 +5,9 @@ import androidx.room.Entity
 import androidx.room.Index
 import androidx.room.PrimaryKey
 
+/**
+ * Expense row. `amount` is stored as **Long cents**.
+ */
 @Entity(tableName = "expenses", indices = [Index("category"), Index("date"), Index("user_id")])
 data class ExpenseEntity(
     @PrimaryKey
@@ -12,7 +15,8 @@ data class ExpenseEntity(
 
     val category: String,
 
-    val amount: Double = 0.0,
+    /** Cents. 1234 = $12.34. */
+    val amount: Long = 0L,
 
     val description: String? = null,
